+++ src/main/webapp/components/index-cards/index.js
